// src/types/species.ts
export interface Species{
    'name': string;
    'classification': string;
    'designation': string;
    'average_height': string;
    'skin_colors': string;
    'hair_colors': string;
    'eye_colors': string;
    'average_lifespan': string;
    'language': string;
    // we use this as a key for now
    'url': string;
}